# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hiberus Order Management API
  description: API REST para gestión de pedidos y pagos (Prueba Técnica PHP/Symfony)
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local development

security:
  - SimulatedAuth: []

paths:
  /api/products:
    get:
      summary: Listar productos (UC-P01)
      description: Obtiene listado paginado con filtros simples
      security: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Buscar por nombre, descripción o categoría
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: sort
          in: query
          schema:
            type: string
            enum: [id, name, price, stock, category]
            default: id
        - name: direction
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
      responses:
        '200':
          description: Lista paginada de productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      summary: Crear producto (UC-P02, solo Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, stock, category]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 255
                description:
                  type: string
                price:
                  type: number
                  minimum: 0.01
                stock:
                  type: integer
                  minimum: 0
                category:
                  type: string
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Error de validación
        '403':
          description: Acceso denegado (no Admin)

  /api/orders:
    post:
      summary: Crear pedido (UC-O01)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId:
                        type: integer
                      quantity:
                        type: integer
                        minimum: 1
      responses:
        '201':
          description: Pedido creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Error de validación o stock insuficiente
        '404':
          description: Producto no encontrado

  /api/orders/{id}:
    get:
      summary: Ver detalle de pedido (UC-O02)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Pedido no pertenece al usuario
        '404':
          description: Pedido no encontrado

  /api/orders/{id}/checkout:
    post:
      summary: Checkout simulado (UC-O03)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pago exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  order:
                    $ref: '#/components/schemas/Order'
        '402':
          description: Pago fallido (80% probabilidad de éxito)
        '400':
          description: Pedido ya procesado

components:
  securitySchemes:
    SimulatedAuth:
      type: apiKey
      in: header
      name: X-User-Id
      description: |
        Autenticación simulada.Enviar headers:
        - X-User-Id: ID del usuario (ej: 1)
        - X-User-Role: ADMIN o CLIENT

  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        stock:
          type: integer
        category:
          type: string
        createdAt:
          type: string
          format: date-time

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number

    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        status:
          type: string
          enum: [pending, paid, failed]
        total:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true